Moved for stuff to string.cc
git-svn-id: file:///Users/tack/GecodeGitMigration/gecode-svn-mirror/gecode/trunk@4903 e85b7adc-8362-4630-8c63-7469d557c915

+++ gecode/support/dynamic-array.hh
@@ -115,7 +115,7 @@ namespace Gecode { namespace Support {
   void
   DynamicArray<T>::resize(int i) {
     int m = std::max(i+1, (3*n)/2);
-    x = Memory::brealloc(x,n,m);
+    x = Memory::brealloc(x,static_cast<size_t>(n),static_cast<size_t>(m));
     n = m;
   }
